Created dictionaries for the effected SNP height means, and standard deviation buffs.
Created a check system to see if the SNP is in the effected list and then to check if the correct A/B Value is present.

Future: Work on a check system to see if an environment effect is present (low water, etc).

Clean up code for readability and add more functions.